fix: move utility function into helpers
